Add support for zaplib.callRustInSameThreadSync in JS main thread
I did this by initializing our app in the browser’s main thread, which
generally works with things like Mutexes and such, since there is no
possibility for contention yet. We then “transfer ownership” of `Cx`
(implicitly) to the Rust worker, which uses it from there on out.

For the `Cx::call_rust_in_same_thread_sync_fn`, we can’t use a `RwLock`
though, since it needs to be accessed in the main JS thread whenever
`zaplib.callRustInSameThreadSync` gets called. Which is not allowed for
a `RwLock` since it can block. Therefore I changed it so that you’re
only allowed to set it during initialization. Then we can safely read it
in any thread we need it afterwards.

I also updated the compatibility tables in the docs to reflect this new
state of things.

This also paves the way for things like pure single-threaded usage of
Zaplib (#65), or even doing the thread-based WebGL rendering fully in
user-space (no ticket yet).