Add Github Action to build and push to Docker
Using the previous image as the cache here was key, this makes the build
really fast.

For now, we don’t ever bust the cache. If that is important, we should
on occasion do a manual build. We can worry more about this later, for
now I just want to have at least something happen when we change the
Dockerfile. I added a TODO about this.