Use `globalThis` instead of `self`
`self` works in the browser main thread and in workers. But it does not
work in NodeJS. :( `globalThis` was introduced as a variable that is
truly supported in all environments to represent the global context.

This makes our code more compatible with NodeJS.

I also renamed unrelated uses of `self` to `zelf` as to not overload the
aforementioned `self`. I added a comment in the places where we do this.