Also call `onRenderingPanic` for internal RPC errors
Most of these are related to rendering also (mouse events etc), and they
should also never happen so we should assume that the application gets
into an unrecoverable state when they happen.

I also made the default `onRenderingPanic` call throw an uncaught error
so it actually fails tests. See comment in the code for more details.